Refactor About module to use suspend functions
The `AboutRepository` and `DefaultAboutRepository` now use a suspend function `getAboutInfoStream()` instead of a Flow to retrieve about screen data.

The `AboutViewModel` has been updated to call this suspend function and handle potential errors using `runCatching`. It also wraps the call in a `flow` builder to maintain the reactive stream for the UI.